+++ AppCenter/AppCenter/Internals/Storage/MSDBStorage.m
@@ -111,12 +111,12 @@
   NSString *query =
       [NSString stringWithFormat:@"SELECT COUNT(*) FROM \"sqlite_master\" WHERE \"type\"='table' AND \"name\"='%@';", tableName];
   NSArray<NSArray *> *result = [MSDBStorage executeSelectionQuery:query inOpenedDatabase:db];
-  return (result.count > 0) ? [(NSNumber *)result[0][0] boolValue] : NO;
+  return result.count > 0 && result[0].count > 0 ? [(NSNumber *)result[0][0] boolValue] : NO;
 }
 
 + (NSUInteger)versionInOpenedDatabase:(void *)db {
   NSArray<NSArray *> *result = [MSDBStorage executeSelectionQuery:@"PRAGMA user_version" inOpenedDatabase:db];
-  return (result.count > 0) ? [(NSNumber *)result[0][0] unsignedIntegerValue] : 0;
+  return result.count > 0 && result[0].count > 0 ? [(NSNumber *)result[0][0] unsignedIntegerValue] : 0;
 }
 
 + (void)setVersion:(NSUInteger)version inOpenedDatabase:(void *)db {
@@ -127,7 +127,7 @@
 + (void)enableAutoVacuumInOpenedDatabase:(void *)db {
   NSArray<NSArray *> *result = [MSDBStorage executeSelectionQuery:@"PRAGMA auto_vacuum" inOpenedDatabase:db];
   int vacuumMode = 0;
-  if ([result count] != 0){
+  if (result.count > 0 && result[0].count > 0){
     vacuumMode = [(NSNumber *)result[0][0] intValue];
   }
   BOOL autoVacuumDisabled = vacuumMode != 1;
@@ -292,17 +292,17 @@
 
 + (long)getPageSizeInOpenedDatabase:(void *)db {
   NSArray<NSArray *> *rows = [MSDBStorage executeSelectionQuery:@"PRAGMA page_size;" inOpenedDatabase:db];
-  return rows.count != 0 ? [(NSNumber *)rows[0][0] longValue] : 0;
+  return rows.count > 0 && rows[0].count > 0 ? [(NSNumber *)rows[0][0] longValue] : 0;
 }
 
 + (long)getPageCountInOpenedDatabase:(void *)db {
   NSArray<NSArray *> *rows = [MSDBStorage executeSelectionQuery:@"PRAGMA page_count;" inOpenedDatabase:db];
-  return rows.count != 0 ? [(NSNumber *)rows[0][0] longValue] : 0;
+  return rows.count > 0 && rows[0].count > 0 ? [(NSNumber *)rows[0][0] longValue] : 0;
 }
 
 + (long)getMaxPageCountInOpenedDatabase:(void *)db {
   NSArray<NSArray *> *rows = [MSDBStorage executeSelectionQuery:@"PRAGMA max_page_count;" inOpenedDatabase:db];
-  return rows.count != 0 ? [(NSNumber *)rows[0][0] longValue] : 0;
+  return rows.count > 0 && rows[0].count > 0 ? [(NSNumber *)rows[0][0] longValue] : 0;
 }
 
 + (int)setMaxPageCount:(long)maxPageCount inOpenedDatabase:(void *)db {